requirements.txt
discord
boto3
awscli
requests
python-dotenv
